move third_party && fix the bug of install
