+++ MANIFEST.in
@@ -3,3 +3,4 @@ include README.md
 include LICENSE
 include bmwhisper/assets/*
 include bmwhisper/normalizers/english.json
+include bmwhisper/third_party/untpu/lib/libuntpu.so
